Refactor: `PermissionManagementDialog` code structure
