+++ src/modules/dashboard/components/PermissionManagementDialog/PermissionManagementDialog.tsx
@@ -1,5 +1,5 @@
 import { FormikProps } from 'formik';
-import React, { useCallback, useState, useMemo } from 'react';
+import React, { useCallback, useState } from 'react';
 import * as yup from 'yup';
 import { defineMessages, FormattedMessage } from 'react-intl';
 import { ROOT_DOMAIN_ID, ColonyRole } from '@colony/colony-js';
@@ -15,8 +15,6 @@ import {
 } from '~utils/actions';
 import { ActionTypes } from '~redux/index';
 import { useTransformer, WizardDialogType } from '~utils/hooks';
-import { ItemDataType } from '~core/OmniPicker';
-import Heading from '~core/Heading';
 import Button from '~core/Button';
 import PermissionsLabel from '~core/PermissionsLabel';
 import Dialog, {
@@ -26,55 +24,49 @@ import Dialog, {
 } from '~core/Dialog';
 import { ActionForm } from '~core/Fields';
 import { SpinnerLoader } from '~core/Preloaders';
-import SingleUserPicker, { filterUserSelection } from '~core/SingleUserPicker';
-import PermissionRequiredInfo from '~core/PermissionRequiredInfo';
-import HookedUserAvatar from '~users/HookedUserAvatar';
-import {
-  useLoggedInUser,
-  useColonySubscribedUsersQuery,
-  useUser,
-  AnyUser,
-} from '~data/index';
-import Toggle from '~core/Fields/Toggle';
+import { useLoggedInUser, useUser, AnyUser } from '~data/index';
 import NotEnoughReputation from '~dashboard/NotEnoughReputation';
 import { useDialogActionPermissions } from '~utils/hooks/useDialogActionPermissions';
 
 import {
   getUserRolesForDomain,
   getAllRootAccounts,
-  getAllUserRolesForDomain,
 } from '../../../transformers';
-import PermissionManagementForm from './PermissionManagementForm';
 import { availableRoles } from './constants';
+import { Address } from '~types/index';
+
+import PermissionManagementForm from './PermissionManagementForm';
 
 import styles from './PermissionManagementDialog.css';
+
+const validationSchema = yup.object().shape({
+  domainId: yup.number().required(),
+  user: yup.object().required(),
+  roles: yup.array().ensure(),
+  annotation: yup.string().max(4000),
+  forceAction: yup.boolean(),
+});
 
 const displayName = 'dashboard.PermissionManagementDialog';
 
 const MSG = defineMessages({
-  title: {
-    id: 'dashboard.PermissionManagementDialog.title',
-    defaultMessage: 'Permissions',
-  },
-  selectUser: {
-    id: 'dashboard.PermissionManagementDialog.selectUser',
-    defaultMessage: 'Member',
-  },
   noPermissionFrom: {
     id: 'dashboard.PermissionManagementDialog.noPermissionFrom',
     defaultMessage: `You do not have the {roleRequired} permission required to take this action.`,
   },
 });
 
+interface FormValues {
+  domainId: string;
+  user: Address;
+  roles: string[];
+  annotation: string;
+  forceAction: boolean;
+}
+
 type Props = DialogProps &
   Partial<WizardDialogType<object>> &
   ActionDialogProps;
-
-const UserAvatar = HookedUserAvatar({ fetchUser: false });
-
-const supRenderAvatar = (address: string, item: ItemDataType<AnyUser>) => (
-  <UserAvatar address={address} user={item} size="xs" notSet={false} />
-);
 
 const PermissionManagementDialog = ({
   colony: { colonyAddress, colonyName, domains },
@@ -108,14 +100,6 @@ const PermissionManagementDialog = ({
     ROOT_DOMAIN_ID,
   );
 
-  const { data: colonySubscribedUsers } = useColonySubscribedUsersQuery({
-    variables: {
-      colonyAddress,
-    },
-  });
-
-  const subscribedUsers = colonySubscribedUsers?.subscribedUsers || [];
-
   const currentUserRoles = useTransformer(getUserRolesForDomain, [
     colony,
     // CURRENT USER!
@@ -144,17 +128,6 @@ const PermissionManagementDialog = ({
     selectedDomainId,
   ]);
 
-  const domainRoles = useTransformer(getAllUserRolesForDomain, [
-    colony,
-    selectedDomainId,
-  ]);
-
-  const directDomainRoles = useTransformer(getAllUserRolesForDomain, [
-    colony,
-    selectedDomainId,
-    true,
-  ]);
-
   const rootAccounts = useTransformer(getAllRootAccounts, [colony]);
 
   const transform = useCallback(
@@ -181,54 +154,14 @@ const PermissionManagementDialog = ({
     [colonyAddress, selectedDomainId],
   );
 
-  const domainRolesArray = useMemo(
-    () =>
-      domainRoles
-        .sort(({ roles }) => (roles.includes(ColonyRole.Root) ? -1 : 1))
-        .filter(({ roles }) => !!roles.length)
-        .map(({ address, roles }) => {
-          const directUserRoles = directDomainRoles.find(
-            ({ address: userAddress }) => userAddress === address,
-          );
-          return {
-            userAddress: address,
-            roles,
-            directRoles: directUserRoles ? directUserRoles.roles : [],
-          };
-        }),
-    [directDomainRoles, domainRoles],
-  );
-
-  const validationSchema = yup.object().shape({
-    domainId: yup.number().required(),
-    user: yup.object().required(),
-    roles: yup.array().ensure(),
-    annotation: yup.string().max(4000),
-  });
-
   const domain = domains?.find(
     ({ ethDomainId }) => ethDomainId === selectedDomainId,
   );
-
-  const members = subscribedUsers.map((user) => {
-    const {
-      profile: { walletAddress },
-    } = user;
-    const domainRole = domainRolesArray.find(
-      (rolesObject) => rolesObject.userAddress === walletAddress,
-    );
-    return {
-      ...user,
-      roles: domainRole ? domainRole.roles : [],
-      directRoles: domainRole ? domainRole.directRoles : [],
-    };
-  });
 
   const canEditPermissions =
     (selectedDomainId === ROOT_DOMAIN_ID &&
       currentUserRolesInRoot.includes(ColonyRole.Root)) ||
     currentUserRolesInRoot.includes(ColonyRole.Architecture);
-  const requiredRoles: ColonyRole[] = [ColonyRole.Architecture];
 
   const [userHasPermission, onlyForceAction] = useDialogActionPermissions(
     colony.colonyAddress,
@@ -260,62 +193,28 @@ const PermissionManagementDialog = ({
           success={getFormAction('SUCCESS')}
           transform={transform}
         >
-          {({
-            isSubmitting,
-            isValid,
-            initialValues,
-            values,
-          }: FormikProps<any>) => {
+          {(formValues: FormikProps<FormValues>) => {
+            const { values, isSubmitting, isValid, initialValues } = formValues;
             if (values.forceAction !== isForce) {
               setIsForce(values.forceAction);
             }
             return (
               <div className={styles.dialogContainer}>
-                <DialogSection appearance={{ theme: 'heading' }}>
-                  <Heading
-                    appearance={{
-                      size: 'medium',
-                      margin: 'none',
-                      theme: 'dark',
-                    }}
-                    text={MSG.title}
-                    textValues={{ domain: domain && domain.name }}
-                  />
-                  {canEditPermissions && isVotingExtensionEnabled && (
-                    <Toggle label={{ id: 'label.force' }} name="forceAction" />
-                  )}
-                </DialogSection>
-                {!userHasPermission && (
-                  <DialogSection>
-                    <PermissionRequiredInfo requiredRoles={requiredRoles} />
-                  </DialogSection>
-                )}
-                <DialogSection appearance={{ theme: 'sidePadding' }}>
-                  <div className={styles.singleUserContainer}>
-                    <SingleUserPicker
-                      data={members}
-                      label={MSG.selectUser}
-                      name="user"
-                      filter={filterUserSelection}
-                      onSelected={setSelectedUser}
-                      renderAvatar={supRenderAvatar}
-                      disabled={inputDisabled}
-                    />
-                  </div>
-                </DialogSection>
-                <DialogSection appearance={{ theme: 'sidePadding' }}>
-                  <PermissionManagementForm
-                    currentUserRoles={currentUserRoles}
-                    domainId={selectedDomainId}
-                    rootAccounts={rootAccounts}
-                    userDirectRoles={userDirectRoles}
-                    currentUserRolesInRoot={currentUserRolesInRoot}
-                    userInheritedRoles={userInheritedRoles}
-                    colonyDomains={domains}
-                    onDomainSelected={setSelectedDomainId}
-                    inputDisabled={inputDisabled}
-                  />
-                </DialogSection>
+                <PermissionManagementForm
+                  {...formValues}
+                  colony={colony}
+                  currentUserRoles={currentUserRoles}
+                  domainId={selectedDomainId}
+                  rootAccounts={rootAccounts}
+                  userDirectRoles={userDirectRoles}
+                  currentUserRolesInRoot={currentUserRolesInRoot}
+                  userInheritedRoles={userInheritedRoles}
+                  onDomainSelected={setSelectedDomainId}
+                  onChangeSelectedUser={setSelectedUser}
+                  inputDisabled={inputDisabled}
+                  userHasPermission={userHasPermission}
+                  isVotingExtensionEnabled={isVotingExtensionEnabled}
+                />
                 {!userHasPermission && (
                   <DialogSection appearance={{ theme: 'sidePadding' }}>
                     <div className={styles.noPermissionFromMessage}>
