+++ src/modules/dashboard/components/PermissionManagementDialog/PermissionManagementForm.tsx
@@ -3,16 +3,30 @@ import { defineMessages } from 'react-intl';
 import { ColonyRole, ROOT_DOMAIN_ID } from '@colony/colony-js';
 import sortBy from 'lodash/sortBy';
 
-import { Address } from '~types/index';
-import { DomainFieldsFragment } from '~data/generated';
 import { InputLabel, Select, Annotations } from '~core/Fields';
+import { DialogSection } from '~core/Dialog';
+import Heading from '~core/Heading';
+import PermissionRequiredInfo from '~core/PermissionRequiredInfo';
+import SingleUserPicker, { filterUserSelection } from '~core/SingleUserPicker';
+import Toggle from '~core/Fields/Toggle';
+import { ItemDataType } from '~core/OmniPicker';
+import HookedUserAvatar from '~users/HookedUserAvatar';
+
+import { Address } from '~types/index';
+import { useColonySubscribedUsersQuery, AnyUser, Colony } from '~data/index';
+import { useTransformer } from '~utils/hooks';
+import { getAllUserRolesForDomain } from '../../../transformers';
+import { availableRoles } from './constants';
 
 import PermissionManagementCheckbox from './PermissionManagementCheckbox';
-import { availableRoles } from './constants';
 
 import styles from './PermissionManagementDialog.css';
 
 const MSG = defineMessages({
+  title: {
+    id: 'dashboard.PermissionManagementDialog.PermissionManagementForm.title',
+    defaultMessage: 'Permissions',
+  },
   domain: {
     id: 'dashboard.PermissionManagementDialog.PermissionManagementForm.domain',
     defaultMessage: 'Team',
@@ -23,36 +37,59 @@ const MSG = defineMessages({
     defaultMessage: 'Permissions',
   },
   annotation: {
-    id:
-      // eslint-disable-next-line max-len
-      'dashboard.PermissionManagementDialog.PermissionManagementForm.annotation',
+    id: `dashboard.PermissionManagementDialog.PermissionManagementForm.annotation`,
     defaultMessage: 'Explain why you’re making these changes (optional)',
+  },
+  selectUser: {
+    id: `dashboard.PermissionManagementDialog.PermissionManagementForm.selectUser`,
+    defaultMessage: 'Member',
   },
 });
 
 interface Props {
+  colony: Colony;
   currentUserRoles: ColonyRole[];
   domainId: number;
   rootAccounts: Address[];
   userDirectRoles: ColonyRole[];
   currentUserRolesInRoot: ColonyRole[];
   userInheritedRoles: ColonyRole[];
-  colonyDomains: DomainFieldsFragment[];
   onDomainSelected: (domain: number) => void;
+  onChangeSelectedUser: (user: AnyUser) => void;
   inputDisabled: boolean;
+  userHasPermission: boolean;
+  isVotingExtensionEnabled: boolean;
 }
 
+const UserAvatar = HookedUserAvatar({ fetchUser: false });
+
+const supRenderAvatar = (address: string, item: ItemDataType<AnyUser>) => (
+  <UserAvatar address={address} user={item} size="xs" notSet={false} />
+);
+
 const PermissionManagementForm = ({
+  colony: { colonyAddress, domains },
+  colony,
   currentUserRoles,
   domainId,
   rootAccounts,
   userDirectRoles,
   userInheritedRoles,
-  colonyDomains,
-  onDomainSelected,
   currentUserRolesInRoot,
   inputDisabled,
+  userHasPermission,
+  isVotingExtensionEnabled,
+  onDomainSelected,
+  onChangeSelectedUser,
 }: Props) => {
+  const { data: colonySubscribedUsers } = useColonySubscribedUsersQuery({
+    variables: {
+      colonyAddress,
+    },
+  });
+
+  const domain = domains?.find(({ ethDomainId }) => ethDomainId === domainId);
+
   const canSetPermissionsInRoot =
     domainId === ROOT_DOMAIN_ID &&
     currentUserRoles.includes(ColonyRole.Root) &&
@@ -61,6 +98,10 @@ const PermissionManagementForm = ({
   const hasArchitectureInRoot = currentUserRolesInRoot.includes(
     ColonyRole.Architecture,
   );
+  const canEditPermissions =
+    (domainId === ROOT_DOMAIN_ID &&
+      currentUserRolesInRoot.includes(ColonyRole.Root)) ||
+    currentUserRolesInRoot.includes(ColonyRole.Architecture);
 
   // Check which roles the current user is allowed to set in this domain
   const canRoleBeSet = useCallback(
@@ -93,13 +134,58 @@ const PermissionManagementForm = ({
     [domainId, canSetPermissionsInRoot, hasArchitectureInRoot, hasRoot],
   );
 
+  const domainRoles = useTransformer(getAllUserRolesForDomain, [
+    colony,
+    domainId,
+  ]);
+
+  const directDomainRoles = useTransformer(getAllUserRolesForDomain, [
+    colony,
+    domainId,
+    true,
+  ]);
+
   const domainSelectOptions = sortBy(
-    colonyDomains.map(({ ethDomainId, name }) => ({
+    domains.map(({ ethDomainId, name }) => ({
       value: ethDomainId.toString(),
       label: name,
     })),
     ['value'],
   );
+
+  const domainRolesArray = useMemo(
+    () =>
+      domainRoles
+        .sort(({ roles }) => (roles.includes(ColonyRole.Root) ? -1 : 1))
+        .filter(({ roles }) => !!roles.length)
+        .map(({ address, roles }) => {
+          const directUserRoles = directDomainRoles.find(
+            ({ address: userAddress }) => userAddress === address,
+          );
+          return {
+            userAddress: address,
+            roles,
+            directRoles: directUserRoles ? directUserRoles.roles : [],
+          };
+        }),
+    [directDomainRoles, domainRoles],
+  );
+
+  const subscribedUsers = colonySubscribedUsers?.subscribedUsers || [];
+
+  const members = subscribedUsers.map((user) => {
+    const {
+      profile: { walletAddress },
+    } = user;
+    const domainRole = domainRolesArray.find(
+      (rolesObject) => rolesObject.userAddress === walletAddress,
+    );
+    return {
+      ...user,
+      roles: domainRole ? domainRole.roles : [],
+      directRoles: domainRole ? domainRole.directRoles : [],
+    };
+  });
 
   const handleDomainChange = useCallback(
     (value: string) => onDomainSelected(Number(value)),
@@ -116,43 +202,80 @@ const PermissionManagementForm = ({
     [domainId],
   );
 
+  const requiredRoles: ColonyRole[] = [ColonyRole.Architecture];
+
   return (
     <>
-      <div className={styles.domainSelectContainer}>
-        <Select
-          options={domainSelectOptions}
-          label={MSG.domain}
-          name="domainId"
-          appearance={{ theme: 'grey' }}
-          onChange={handleDomainChange}
+      <DialogSection appearance={{ theme: 'heading' }}>
+        <Heading
+          appearance={{
+            size: 'medium',
+            margin: 'none',
+            theme: 'dark',
+          }}
+          text={MSG.title}
+          textValues={{ domain: domain?.name }}
+        />
+        {canEditPermissions && isVotingExtensionEnabled && (
+          <Toggle label={{ id: 'label.force' }} name="forceAction" />
+        )}
+      </DialogSection>
+      {!userHasPermission && (
+        <DialogSection>
+          <PermissionRequiredInfo requiredRoles={requiredRoles} />
+        </DialogSection>
+      )}
+      <DialogSection appearance={{ theme: 'sidePadding' }}>
+        <div className={styles.singleUserContainer}>
+          <SingleUserPicker
+            data={members}
+            label={MSG.selectUser}
+            name="user"
+            filter={filterUserSelection}
+            onSelected={onChangeSelectedUser}
+            renderAvatar={supRenderAvatar}
+            disabled={inputDisabled}
+          />
+        </div>
+      </DialogSection>
+      <DialogSection appearance={{ theme: 'sidePadding' }}>
+        <div className={styles.domainSelectContainer}>
+          <Select
+            options={domainSelectOptions}
+            label={MSG.domain}
+            name="domainId"
+            appearance={{ theme: 'grey' }}
+            onChange={handleDomainChange}
+          />
+        </div>
+        <InputLabel
+          label={MSG.permissionsLabel}
+          appearance={{ colorSchema: 'grey' }}
+        />
+        <div className={styles.permissionChoiceContainer}>
+          {filteredRoles.map((role) => {
+            const roleIsInherited =
+              !userDirectRoles.includes(role) &&
+              userInheritedRoles.includes(role);
+            return (
+              <PermissionManagementCheckbox
+                key={role}
+                disabled={
+                  inputDisabled || !canRoleBeSet(role) || roleIsInherited
+                }
+                role={role}
+                asterisk={roleIsInherited}
+                domainId={domainId}
+              />
+            );
+          })}
+        </div>
+        <Annotations
+          label={MSG.annotation}
+          name="annotationMessage"
           disabled={inputDisabled}
         />
-      </div>
-      <InputLabel
-        label={MSG.permissionsLabel}
-        appearance={{ colorSchema: 'grey' }}
-      />
-      <div className={styles.permissionChoiceContainer}>
-        {filteredRoles.map((role) => {
-          const roleIsInherited =
-            !userDirectRoles.includes(role) &&
-            userInheritedRoles.includes(role);
-          return (
-            <PermissionManagementCheckbox
-              key={role}
-              disabled={inputDisabled || !canRoleBeSet(role) || roleIsInherited}
-              role={role}
-              asterisk={roleIsInherited}
-              domainId={domainId}
-            />
-          );
-        })}
-      </div>
-      <Annotations
-        label={MSG.annotation}
-        name="annotationMessage"
-        disabled={inputDisabled}
-      />
+      </DialogSection>
     </>
   );
 };
